refactor: place 도메인 컨트롤러 통합
