+++ src/main/java/com/babyblackdog/ddogdog/place/controller/PlaceRestController.java
@@ -1,9 +1,14 @@
-package com.babyblackdog.ddogdog.place.room.controller;
+package com.babyblackdog.ddogdog.place.controller;
 
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 
 import com.babyblackdog.ddogdog.common.date.StayPeriod;
 import com.babyblackdog.ddogdog.common.date.TimeProvider;
+import com.babyblackdog.ddogdog.place.hotel.controller.dto.HotelResponse;
+import com.babyblackdog.ddogdog.place.hotel.controller.dto.HotelResponses;
+import com.babyblackdog.ddogdog.place.hotel.model.vo.Province;
+import com.babyblackdog.ddogdog.place.hotel.service.HotelService;
+import com.babyblackdog.ddogdog.place.hotel.service.dto.HotelResult;
 import com.babyblackdog.ddogdog.place.room.controller.dto.RoomResponse;
 import com.babyblackdog.ddogdog.place.room.controller.dto.RoomResponses;
 import com.babyblackdog.ddogdog.place.room.service.RoomService;
@@ -20,19 +25,42 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping(path = "/rooms", produces = APPLICATION_JSON_VALUE)
-public class RoomRestController {
+@RequestMapping(path = "/places", produces = APPLICATION_JSON_VALUE)
+class PlaceRestController {
 
+  private final HotelService hotelService;
   private final RoomService roomService;
   private final TimeProvider timeProvider;
 
-  public RoomRestController(RoomService roomService, TimeProvider timeProvider) {
+  PlaceRestController(HotelService hotelService, RoomService roomService, TimeProvider timeProvider) {
+    this.hotelService = hotelService;
     this.roomService = roomService;
     this.timeProvider = timeProvider;
   }
 
-  @GetMapping(path = "/{roomId}")
+  // GET /places/{hotelId}
+  @GetMapping("/{hotelId}")
+  public ResponseEntity<HotelResponse> getHotel(@PathVariable Long hotelId) {
+    HotelResult result = hotelService.findHotelById(hotelId);
+    return ResponseEntity
+        .status(HttpStatus.OK)
+        .body(HotelResponse.of(result));
+  }
+
+  // GET /places?province=강원
+  @GetMapping
+  public ResponseEntity<HotelResponses> getHotelsInProvince(@RequestParam String province,
+      Pageable pageable) {
+    Page<HotelResult> result = hotelService.findHotelsInProvince(new Province(province), pageable);
+    return ResponseEntity
+        .status(HttpStatus.OK)
+        .body(HotelResponses.of(result));
+  }
+
+  // GET /places/{hotelId}/{roomId}?checkIn=&checkOut=
+  @GetMapping(path = "/{hotelId}/{roomId}")
   public ResponseEntity<RoomResponse> getRoomForDuration(
+      @PathVariable Long hotelId,
       @PathVariable Long roomId,
       @RequestParam LocalDate checkIn,
       @RequestParam LocalDate checkOut
@@ -44,9 +72,10 @@ public class RoomRestController {
         .body(RoomResponse.of(roomResult));
   }
 
-  @GetMapping
+  // GET /places/{hotelId}/rooms?checkIn=&checkOut=
+  @GetMapping(path = "/{hotelId}/rooms")
   public ResponseEntity<RoomResponses> getRoomsOfHotelForDuration(
-      @RequestParam Long hotelId,
+      @PathVariable Long hotelId,
       @RequestParam LocalDate checkIn,
       @RequestParam LocalDate checkOut,
       Pageable pageable
